Rename http/http_decoding to http/response_decoding.
