+++ packages/hurl/src/http/mod.rs
@@ -30,7 +30,7 @@ pub use self::response::{Response, Version};
 pub use self::version::libcurl_version_info;
 
 mod client;
-mod content_decoding;
+mod response_decoding;
 mod core;
 mod error;
 mod options;
